pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wimbledon_simulator"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
description = "A sophisticated tennis tournament simulator using multi-agent architecture"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=0.24.2",
    "openai>=1.0.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=5.4.1",
    "rich>=10.12.0",
    "chromadb>=0.3.0",
    "pytest>=6.2.5",
    "pytest-cov>=2.12.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "pytest-cov>=2.12.1"
] 